backend/requirements.txt
fastapi
uvicorn

numpy
pandas
requests
python-dotenv
pytesseract
opencv-python
beautifulsoup4
openpyxl

tensorflow
torch
